HW3 ex2

+++ server/server_constants.go
@@ -7,7 +7,7 @@ const CHUNK_SIZE = 8 * 1024
 const PACKET_SIZE = CHUNK_SIZE + 1024
 const SHA_SIZE = 32
 
-const HOP_LIMIT = 10
+const HOP_LIMIT = uint32(10)
 
 
 const STATUS_COUNTDOWN_TIME = 10
